+++ includes/raytracing.h
@@ -19,18 +19,18 @@
 /*
 ** Intersect
 */
-float		intersect_sphere(t_vector origin, t_vector direction,
+double		intersect_sphere(t_vector origin, t_vector direction,
 t_sphere *object);
-float		intersect_plan(t_vector origin, t_vector direction,
+double		intersect_plan(t_vector origin, t_vector direction,
 t_plan *object);
 
 /*
 ** Raytracing
 */
 int			trace_ray(t_vector direction, t_scene *scene);
-float		closest_intersection(t_vector origin, t_vector direction,
+double		closest_intersection(t_vector origin, t_vector direction,
 t_scene *scene, t_lstobjects **closest_object);
 int			setup_l_vectors_and_calculate(t_lstobjects *closest_object,
-t_vector direction, float closest_t, t_scene *scene);
+t_vector direction, double closest_t, t_scene *scene);
 
 #endif
