+++ includes/utils.h
@@ -16,7 +16,7 @@
 # include "vector.h"
 
 int		ft_atoi(char *line, int *nb);
-int		ft_atof(char *line, float *nb);
+int		ft_atof(char *line, double *nb);
 int		ft_atoc(char *line, t_vector **color);
 int		ft_atov(char *line, t_vector **vector);
 int		ft_strcmp(const char *s1, const char *s2);
