+++ srcs/scene/light.c
@@ -13,7 +13,7 @@
 #include "error.h"
 #include "light.h"
 
-t_light			*new_default_light(int type, float intensity, t_vector *color)
+t_light			*new_default_light(int type, double intensity, t_vector *color)
 {
 	t_light		*light;
 
@@ -26,7 +26,7 @@ t_light			*new_default_light(int type, float intensity, t_vector *color)
 	return (light);
 }
 
-t_light			*new_point_light(t_vector *pos, float intensity,
+t_light			*new_point_light(t_vector *pos, double intensity,
 t_vector *color)
 {
 	t_light		*light;
